requirements.txt
numpy
python-docx
tkinter
